Split in different files Improve spacing Add `default` for `JOEmojiableConfig
