+++ Pod/Classes/JOEmojiableBtn.swift
@@ -3,33 +3,11 @@
 //  JOEmojiableBtn
 //
 //  Created by Jorge Raul Ovalle Zuleta on 2/28/16.
-//  Copyright © 2016 jorgeovalle. All rights reserved.
+//  Copyright © 2018 Jorge Ovalle. All rights reserved.
 //
 
 import UIKit
 
-/**
- *  Control's configuration, details on Github https://github.com/lojals/JOEmojiableBtn
- */
-public struct JOEmojiableConfig {
-    var spacing: CGFloat
-    var size: CGFloat
-    var minSize: CGFloat
-    var maxSize: CGFloat
-    var s_options_selector: CGFloat
-
-    public init(spacing: CGFloat, size: CGFloat, minSize: CGFloat, maxSize: CGFloat, s_options_selector: CGFloat) {
-        self.spacing  = spacing
-        self.size = size
-        self.minSize = minSize
-        self.maxSize = maxSize
-        self.s_options_selector = s_options_selector
-    }
-}
-
-/**
- *  TODO: use name value to create Option's labels
- */
 public struct JOEmojiableOption {
     var image: String
     var name: String
@@ -69,45 +47,25 @@ open class JOEmojiableBtn: UIButton {
     let s_options_selector: CGFloat
 
     /**
-     Initialization with parameters as default (Facebook reactions iOS App)
-     
-     - parameter frame: Frame of the button will open the selector
-     
-     */
-
-    public override init(frame: CGRect) {
-        self.spacing  = 6
-        self.size = 40
-        self.minSize = 34
-        self.maxSize = 80
-        self.s_options_selector = 30
-        super.init(frame: frame)
-        self.initialize()
-    }
-
-    /**
      Initialization with Custom sizes check Documentation Github project
      
      - parameter frame: Frame of the button will open the selector
      - parameter config: JOEmojiableConfig value with custom sizes
      */
 
-    public init(frame: CGRect, config: JOEmojiableConfig) {
+    public init(frame: CGRect, config: JOEmojiableConfig = .default) {
         self.spacing = config.spacing
         self.size = config.size
         self.minSize = config.minSize
         self.maxSize = config.maxSize
         self.s_options_selector = config.s_options_selector
         super.init(frame: frame)
-        self.initialize()
-    }
-
-    fileprivate func initialize() {
+        
         longTap = UILongPressGestureRecognizer(target: self, action: #selector(JOEmojiableBtn.longTapEvent))
         singleTap = UITapGestureRecognizer(target: self, action: #selector(JOEmojiableBtn.singleTapEvent))
-        self.addGestureRecognizer(longTap)
-        self.addGestureRecognizer(singleTap)
-        self.layer.masksToBounds = false
+        addGestureRecognizer(longTap)
+        addGestureRecognizer(singleTap)
+        layer.masksToBounds = false
         active = false
     }
 
@@ -130,11 +88,11 @@ open class JOEmojiableBtn: UIButton {
                 active = true
                 bgClear = SelectorView(frame: frameSV)
                 bgClear.delegate = self
-                bgClear.backgroundColor = UIColor.clear
+                bgClear.backgroundColor = .clear
 
-                origin = self.superview?.convert(self.frame.origin, to: nil)
+                origin = superview?.convert(self.frame.origin, to: nil)
 
-                if origin != self.frame.origin {
+                if origin != frame.origin {
                     bgClear.frame.origin.x -= origin.x
                     bgClear.frame.origin.y -= origin.y
                 }
@@ -144,9 +102,9 @@ open class JOEmojiableBtn: UIButton {
                 let sizeBtn = CGSize(width: ((CGFloat(dataset.count + 1) * spacing) + (size * CGFloat(dataset.count))), height: size + (2 * spacing))
                 options = UIView(frame: CGRect(x: origin.x, y: origin.y - (sizeBtn.height), width: sizeBtn.width, height: sizeBtn.height))
                 options.layer.cornerRadius  = options.frame.height/2
-                options.backgroundColor     = UIColor.white
+                options.backgroundColor     = .white
                 options.layer.shadowColor   = UIColor.lightGray.cgColor
-                options.layer.shadowOffset  = CGSize(width: 0.0, height: 0.0)
+                options.layer.shadowOffset  = .zero
                 options.layer.shadowOpacity = 0.5
                 options.alpha               = 0.3
                 bgClear.addSubview(options)
@@ -157,11 +115,11 @@ open class JOEmojiableBtn: UIButton {
                 })
 
                 for i in 0..<dataset.count {
-                    let option = UIImageView(frame: CGRect(x: (CGFloat(i+1)*spacing)+(size*CGFloat(i)), y: sizeBtn.height*1.2, width: 10, height: 10))
+                    let option = UIImageView(frame: CGRect(x: (CGFloat(i + 1) * spacing) + (size * CGFloat(i)), y: sizeBtn.height * 1.2, width: 10, height: 10))
                     option.image = UIImage(named: dataset[i].image)
                     option.alpha = 0.6
                     options.addSubview(option)
-                    UIView.animate(withDuration: 0.2, delay: 0.05*Double(i), options: UIViewAnimationOptions.curveEaseInOut, animations: { () -> Void in
+                    UIView.animate(withDuration: 0.2, delay: 0.05*Double(i), options: .curveEaseInOut, animations: { () -> Void in
                         option.frame.origin.y = self.spacing
                         option.alpha = 1
                         option.frame.size = CGSize(width: self.size, height: self.size)
@@ -170,7 +128,7 @@ open class JOEmojiableBtn: UIButton {
                 }
 
                 information = InformationView(frame: CGRect(x: 0, y: origin.y, width: frameSV.width, height: self.frame.height))
-                information.backgroundColor = UIColor.white
+                information.backgroundColor = .white
                 bgClear.addSubview(information)
             } else {
                 print("Please, initialize the dataset.")
@@ -216,10 +174,10 @@ open class JOEmojiableBtn: UIButton {
         information.activateInfo(true)
         UIView.animate(withDuration: 0.3) { () -> Void in
             let sizeBtn = CGSize(width: ((CGFloat(self.dataset.count + 1) * self.spacing) + (self.size * CGFloat(self.dataset.count))), height: self.size + (2 * self.spacing))
-            self.options.frame = CGRect(x: self.origin.x, y: self.origin.y - (self.s_options_selector+sizeBtn.height), width: sizeBtn.width, height: sizeBtn.height)
+            self.options.frame = CGRect(x: self.origin.x, y: self.origin.y - (self.s_options_selector + sizeBtn.height), width: sizeBtn.width, height: sizeBtn.height)
             self.options.layer.cornerRadius = sizeBtn.height / 2
             for (idx, view) in self.options.subviews.enumerated() {
-                view.frame = CGRect(x: (CGFloat(idx + 1) * self.spacing) + (self.size*CGFloat(idx)), y: self.spacing, width: self.size, height: self.size)
+                view.frame = CGRect(x: (CGFloat(idx + 1) * self.spacing) + (self.size * CGFloat(idx)), y: self.spacing, width: self.size, height: self.size)
             }
         }
     }
@@ -258,7 +216,7 @@ open class JOEmojiableBtn: UIButton {
     }
 }
 
-extension JOEmojiableBtn:SelectorViewDelegate {
+extension JOEmojiableBtn: SelectorViewDelegate {
     /**
      Function that track the user's touch, and help to find the selection user want to do.
      
@@ -283,47 +241,5 @@ extension JOEmojiableBtn:SelectorViewDelegate {
         } else {
             self.deActivate(-1)
         }
-    }
-}
-
-open class InformationView: UIView {
-    fileprivate var textInformation: UILabel!
-    open override func draw(_ rect: CGRect) {
-        let dots = UIBezierPath()
-        dots.move(to: CGPoint(x: 18.5, y: (rect.height/2)))
-        dots.addLine(to: CGPoint(x: rect.width, y: (rect.height/2)))
-        dots.lineCapStyle = CGLineCap.round
-        UIColor(red: 0.8, green: 0.81, blue: 0.82, alpha: 1).setStroke()
-        dots.lineWidth = 3
-        let dashes: [CGFloat] = [dots.lineWidth * 0, 37]
-        dots.setLineDash(dashes, count: dashes.count, phase: 0)
-        dots.stroke()
-
-        let lineSuperior = UIBezierPath()
-        lineSuperior.move(to: CGPoint(x: 0, y: 0))
-        lineSuperior.addLine(to: CGPoint(x: rect.width, y: 0))
-        UIColor(red: 0.8, green: 0.81, blue: 0.82, alpha: 1).setStroke()
-        lineSuperior.lineWidth = 1
-        lineSuperior.stroke()
-
-        let lineInferior = UIBezierPath()
-        lineInferior.move(to: CGPoint(x: 0, y: rect.height))
-        lineInferior.addLine(to: CGPoint(x: rect.width, y: rect.height))
-        UIColor(red:0.8, green:0.81, blue:0.82, alpha:1).setStroke()
-        lineInferior.lineWidth = 1
-        lineInferior.stroke()
-
-        textInformation = UILabel(frame: CGRect(x: 0, y: 1, width: rect.width, height: rect.height-2))
-        textInformation.backgroundColor = UIColor.white
-        textInformation.textColor = UIColor(red: 0.57, green: 0.59, blue: 0.64, alpha: 1)
-        textInformation.text = "Release to Cancel"
-        textInformation.textAlignment = .center
-        textInformation.font = UIFont.boldSystemFont(ofSize: 12)
-        textInformation.alpha = 0
-        self.addSubview(textInformation)
-    }
-
-    func activateInfo(_ active: Bool) {
-        textInformation.alpha = active ? 1 : 0
     }
 }
